+++ e2e/e2etests/e2etests.go
@@ -63,7 +63,9 @@ const (
 	TestSolanaDepositName                                 = "solana_deposit"
 	TestSolanaDepositThroughProgramName                   = "solana_deposit_through_program"
 	TestSolanaWithdrawName                                = "solana_withdraw"
+	TestSolanaWithdrawRevertExecutableReceiverName        = "solana_withdraw_revert_executable_receiver"
 	TestSolanaWithdrawAndCallName                         = "solana_withdraw_and_call"
+	TestSolanaWithdrawAndCallInvalidMsgEncodingName       = "solana_withdraw_and_call_invalid_msg_encoding"
 	TestZEVMToSolanaCallName                              = "zevm_to_solana_call"
 	TestSolanaWithdrawAndCallRevertWithCallName           = "solana_withdraw_and_call_revert_with_call"
 	TestSolanaDepositAndCallName                          = "solana_deposit_and_call"
@@ -72,6 +74,7 @@ const (
 	TestSolanaDepositAndCallRevertWithCallThatRevertsName = "solana_deposit_and_call_revert_with_call_that_reverts"
 	TestSolanaDepositAndCallRevertWithDustName            = "solana_deposit_and_call_revert_with_dust"
 	TestSolanaToZEVMCallName                              = "solana_to_zevm_call"
+	TestSolanaToZEVMCallAbortName                         = "solana_to_zevm_call_abort"
 	TestSolanaDepositRestrictedName                       = "solana_deposit_restricted"
 	TestSolanaWithdrawRestrictedName                      = "solana_withdraw_restricted"
 	TestSPLDepositName                                    = "spl_deposit"
@@ -90,6 +93,7 @@ const (
 	TestTONDepositName              = "ton_deposit"
 	TestTONDepositAndCallName       = "ton_deposit_and_call"
 	TestTONDepositAndCallRefundName = "ton_deposit_refund"
+	TestTONDepositRestrictedName    = "ton_deposit_restricted"
 	TestTONWithdrawName             = "ton_withdraw"
 	TestTONWithdrawConcurrentName   = "ton_withdraw_concurrent"
 
@@ -139,6 +143,7 @@ const (
 	TestBitcoinWithdrawInvalidAddressName                  = "bitcoin_withdraw_invalid"
 	TestBitcoinWithdrawRestrictedName                      = "bitcoin_withdraw_restricted"
 	TestBitcoinDepositInvalidMemoRevertName                = "bitcoin_deposit_invalid_memo_revert"
+	TestBitcoinWithdrawRBFName                             = "bitcoin_withdraw_rbf"
 
 	/*
 	 Application tests
@@ -259,6 +264,12 @@ const (
 
 const (
 	CountArgDescription = "count"
+)
+
+// Here are all the dependencies for the e2e tests, add more dependencies here if needed
+var (
+	// DepdencyAllBitcoinDeposits is a dependency to wait for all bitcoin deposit tests to complete
+	DepdencyAllBitcoinDeposits = runner.NewE2EDependency("all_bitcoin_deposits")
 )
 
 // AllE2ETests is an ordered list of all e2e tests
@@ -601,11 +612,27 @@ var AllE2ETests = []runner.E2ETest{
 		runner.WithMinimumVersion("v29.0.0"),
 	),
 	runner.NewE2ETest(
+		TestSolanaWithdrawRevertExecutableReceiverName,
+		"withdraw SOL from ZEVM reverts if executable receiver",
+		[]runner.ArgDefinition{
+			{Description: "amount in lamport", DefaultValue: "1000000"},
+		},
+		TestSolanaWithdrawRevertExecutableReceiver,
+	),
+	runner.NewE2ETest(
 		TestZEVMToSolanaCallName,
 		"call solana program from ZEVM",
 		[]runner.ArgDefinition{},
 		TestZEVMToSolanaCall,
 		runner.WithMinimumVersion("v29.0.0"),
+	),
+	runner.NewE2ETest(
+		TestSolanaWithdrawAndCallInvalidMsgEncodingName,
+		"withdraw SOL from ZEVM and call solana program with invalid msg encoding",
+		[]runner.ArgDefinition{
+			{Description: "amount in lamport", DefaultValue: "1000000"},
+		},
+		TestSolanaWithdrawAndCallInvalidMsgEncoding,
 	),
 	runner.NewE2ETest(
 		TestSolanaWithdrawAndCallRevertWithCallName,
@@ -648,6 +675,13 @@ var AllE2ETests = []runner.E2ETest{
 		"call a zevm contract",
 		[]runner.ArgDefinition{},
 		TestSolanaToZEVMCall,
+		runner.WithMinimumVersion("v30.0.0"),
+	),
+	runner.NewE2ETest(
+		TestSolanaToZEVMCallAbortName,
+		"call a zevm contract and abort",
+		[]runner.ArgDefinition{},
+		TestSolanaToZEVMCallAbort,
 		runner.WithMinimumVersion("v30.0.0"),
 	),
 	runner.NewE2ETest(
@@ -799,6 +833,14 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "amount in nano tons", DefaultValue: "1000000000"}, // 1.0 TON
 		},
 		TestTONDepositAndCallRefund,
+	),
+	runner.NewE2ETest(
+		TestTONDepositRestrictedName,
+		"deposit TON into ZEVM restricted address",
+		[]runner.ArgDefinition{
+			{Description: "amount in nano tons", DefaultValue: "100000000"}, // 0.1 TON
+		},
+		TestTONDepositRestricted,
 	),
 	runner.NewE2ETest(
 		TestTONWithdrawName,
@@ -1137,6 +1179,16 @@ var AllE2ETests = []runner.E2ETest{
 		[]runner.ArgDefinition{},
 		TestBitcoinDepositInvalidMemoRevert,
 		runner.WithMinimumVersion("v29.0.0"),
+	),
+	runner.NewE2ETest(
+		TestBitcoinWithdrawRBFName,
+		"withdraw Bitcoin from ZEVM and replace the outbound using RBF",
+		[]runner.ArgDefinition{
+			{Description: "receiver address", DefaultValue: ""},
+			{Description: "amount in btc", DefaultValue: "0.001"},
+		},
+		TestBitcoinWithdrawRBF,
+		runner.WithDependencies(DepdencyAllBitcoinDeposits),
 	),
 	/*
 	 Application tests