+++ e2e/runner/solana.go
@@ -110,11 +110,13 @@ func (r *E2ERunner) CreateSOLCallInstruction(
 	signer solana.PublicKey,
 	receiver ethcommon.Address,
 	data []byte,
+	revertOptions *solanacontract.RevertOptions,
 ) solana.Instruction {
 	callData, err := borsh.Serialize(solanacontract.CallInstructionParams{
 		Discriminator: solanacontract.DiscriminatorCall,
 		Receiver:      receiver,
 		Memo:          data,
+		RevertOptions: revertOptions,
 	})
 	require.NoError(r, err)
 
@@ -542,6 +544,7 @@ func (r *E2ERunner) SOLCall(
 	signerPrivKey *solana.PrivateKey,
 	receiver ethcommon.Address,
 	data []byte,
+	revertOptions *solanacontract.RevertOptions,
 ) solana.Signature {
 	// if signer is not provided, use the runner account as default
 	if signerPrivKey == nil {
@@ -550,7 +553,7 @@ func (r *E2ERunner) SOLCall(
 	}
 
 	// create 'call' instruction
-	instruction := r.CreateSOLCallInstruction(signerPrivKey.PublicKey(), receiver, data)
+	instruction := r.CreateSOLCallInstruction(signerPrivKey.PublicKey(), receiver, data, revertOptions)
 
 	// create and sign the transaction
 	limit := computebudget.NewSetComputeUnitLimitInstruction(100000).Build() // 100k compute unit limit
@@ -605,7 +608,7 @@ func (r *E2ERunner) WithdrawAndCallSOLZRC20(
 	to solana.PublicKey,
 	amount *big.Int,
 	approveAmount *big.Int,
-	data []byte,
+	msgEncoded []byte,
 	revertOptions gatewayzevm.RevertOptions,
 ) *ethtypes.Transaction {
 	// approve
@@ -613,22 +616,6 @@ func (r *E2ERunner) WithdrawAndCallSOLZRC20(
 	require.NoError(r, err)
 	receipt := utils.MustWaitForTxReceipt(r.Ctx, r.ZEVMClient, tx, r.Logger, r.ReceiptTimeout)
 	utils.RequireTxSuccessful(r, receipt, "approve")
-
-	// create encoded msg
-	connectedPda, err := solanacontract.ComputeConnectedPdaAddress(ConnectedProgramID)
-	require.NoError(r, err)
-	msg := solanacontract.ExecuteMsg{
-		Accounts: []solanacontract.AccountMeta{
-			{PublicKey: [32]byte(connectedPda.Bytes()), IsWritable: true},
-			{PublicKey: [32]byte(r.ComputePdaAddress().Bytes()), IsWritable: false},
-			{PublicKey: [32]byte(r.GetSolanaPrivKey().PublicKey().Bytes()), IsWritable: true},
-			{PublicKey: [32]byte(solana.SystemProgramID.Bytes()), IsWritable: false},
-		},
-		Data: data,
-	}
-
-	msgEncoded, err := msg.Encode()
-	require.NoError(r, err)
 
 	// withdraw
 	tx, err = r.GatewayZEVM.WithdrawAndCall0(